Moved SubscribableRemoteStream into Sherlock namespace and subdirectory.
